Put database and routers on separate packages
